Working on the wendigo spawning mechanics
Wendigo will attempt to spawn in behind the player periodically. If it is not seen it the position it is attempted to be spawned in slowly closes in on the view frustrum until it is seen by the player. Then it initiates the chase mechanic.

Wendigoview handles the FSM for the wendigo's state though that should probably be transferred to wendigo controller possibly.

Wendigo controller handles the actual spawning of the wendigo.